requirements.txt
google-cloud-storage==3.1.0
# numpy==2.0.2
numpy==1.26.4
pandas==2.2.3
polars==1.26.0
psycopg2-binary
python-dotenv==1.1.0
requests==2.32.3
requests-ip-rotator==1.0.14
SQLAlchemy==2.0.40
urllib3==1.26.20
pyspark==3.5.1
